add some basic tests
