+++ Cargo.toml
@@ -1,5 +1,5 @@
 [package]
-name = "stop-on-call"
+name = "stop_on_call"
 version = "0.1.0"
 edition = "2024"
 
@@ -7,3 +7,7 @@ edition = "2024"
 axum = { version = "0.8.1" }
 tokio = { version = "1.43.0", features = ["full"] }
 tokio-util = "0.7.13"
+
+[dev-dependencies]
+tokio = { version = "1", features = ["full"] }
+reqwest = { version = "0.11", features = ["json"] }
